requirements.txt
fastapi==0.109.2
motor==3.3.2
pymongo==4.6.1
python-dotenv==1.0.1
uvicorn==0.27.1
pydantic==2.6.1
pydantic-settings==2.1.0 